Move old files into old directory
