+++ old/template
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<meta charset="UTF-8">
+	<title id="title"></title>
+	<script type="text/javascript" src="https://cdn.jsdelivr.net/npm/clipboard@2/dist/clipboard.min.js"></script>
+	<script type="text/javascript" src="template.js"></script>
+	<script type="text/javascript">
+		//define variables of this page
+		let title = "{title}";
+		let event_ID = {event_ID};
+		let place_order = [{place_order}];
+		//第二個參數 8 是指九月，9 是指十月！
+		let dates = [
+			{dates}
+		];
+	</script>
+</head>
+<body>
+	<h1 id="print_title"></h1>
+	<div id="test">
+		網址範例：<br>https://host.cc.ntu.edu.tw/activities/placeApplyDetail.aspx?From=placeApply.aspx&Place_ID=
+		<select id="testPlace" onchange="onChangeOfTestPlaceID()">
+		</select>
+		&Act_ID=
+		<input type="text" id="testActID" disabled style="width: 50px">
+		&Start=
+		<input type="text" id="testStart" disabled style="width: 50px">
+		&End=
+		<input type="text" id="testEnd" disabled style="width: 50px">
+		&Date=
+		<input type="text" id="testYear" style="width: 40px">/
+		<input type="text" id="testMonth" style="width: 25px">/
+		<input type="text" id="testDay" style="width: 25px">
+		<br>
+		<button id="testButton">點擊複製範例</button>
+	</div>
+	<div id="list">
+	</div>
+	<script type="text/javascript">
+		document.getElementById("title").innerHTML = title;
+		document.getElementById("print_title").innerHTML = "{{0}}借場必勝秘笈".format(title);
+		test_generate(event_ID, place_order, dates);
+		generate(event_ID, place_order, dates);
+	</script>
+</body>
+</html>
